pyproject.toml
[tool.poetry]
name = "dpsprep"
version = "2.2.9"
description = "A DjVu to PDF converter with a focus on small output size and the ability to preserve document outlines and text layers"
authors = ["Kevin Arthur Schiff Croker", "Ianis Vasilev"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.7"
djvulibre-python = "^0.9.0"
fpdf2 = "^2.8.1"
loguru = "^0.7.2"
ocrmypdf = {version = "^15.4.4", optional = true}
pdfrw = "^0.4"
pillow = "^11.0.0"

[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"
pytest = "^7.3.1"
pytest-image-diff = "^0.0.11"
ruff = "^0.7.1"
types-fpdf2 = "^2.8.1.20241011"
types-pillow = "^10.2.0.20240822"

[tool.poetry.extras]
compress = ["ocrmypdf"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "COM812",
  "D1", "D203", "D213",
  "E501",
  "FA100",
  "PLR2004",
  "PTH123",
  "Q000",
  "S324"
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.lint.per-file-ignores]
"test_*.py" = ["S101"]

[[tool.mypy.overrides]]
module = [
  "djvu.*",
  "ocrmypdf.*",
  "pdfrw.*",
  "pytest_image_diff.*"
]
ignore_missing_imports = true
